Move movieInfo, Player, Comments to separate colders
