+++ client/src/components/Movie/Comments/Comments.tsx
@@ -1,0 +1,15 @@
+import { Typography } from '@mui/material';
+import Text from '../../Text';
+
+const Comments = ({ movieId }: { movieId: number }) => {
+	return (
+		<>
+			<Typography variant="h6" color={'white'}>
+				<Text tid="plot" />
+			</Typography>
+			<Typography color={'text.secondary'} textAlign={'left'}></Typography>
+		</>
+	);
+};
+
+export default Comments;
