+++ client/src/components/Movie/Player/Player.reducer.tsx
